working on rework
